requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
flask_sqlalchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
marshmallow==3.22.0
Flasgger==0.9.7.1
prometheus-flask-exporter==0.23.1
gunicorn==22.0.0
waitress==3.0.0
APScheduler==3.10.4
itsdangerous==2.2.0
requests==2.32.3
structlog==24.1.0
pywin32==306; platform_system == "Windows"

Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
